# API - updated/openapi.yaml
openapi: 3.0.0
info:
  title: WDE - Assignment 3.1 -  updated
  description: Recipes website API
  contact:
    name: yarden curiel & aviv amsellem.
    email: curiey@post.bgu.ac.il
  version: 1.1.1
servers:
- url: http://localhost:3000/
...
paths:
...
  /login:
    post:
      tags:
      - user
      summary: login method.
      description: Login and set session id.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body'
        required: true
      responses:
        "200":
          description: Login Successfully. The session ID is returned in a cookie
            named `JSESSIONID`. You need to include this cookie in subsequent requests.
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: JSESSIONID=abcde12345
        "401":
          description: Autorization failed, Username or Password incorrect.
  /register:
    post:
      tags:
      - user
      summary: Registration method.
      description: Registration page for a new user.
      requestBody:
        description: Registration page for a new user.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_1'
        required: true
      responses:
        "200":
          description: OK. registration complete successfully
        "400":
          description: Bad Request. Not all reqired argument was given.
        "401":
          description: User already logged in.
        "409":
          description: Username is already taken.
  /logout:
    get:
      tags:
      - user
      summary: User log out from website.
      description: User Request to logout from the website.
      responses:
        "200":
          description: logout succeeded.
        "401":
          description: must logged in before logging out.
      security:
      - cookieAuth: []
...
  /createRecipe:
    post:
      tags:
      - profile
      summary: Upload a new recipe.
      description: user upload a new content to the webiste.
      requestBody:
        description: New recipe that the user want to upload.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseRecipe'
        required: true
      responses:
        "200":
          description: OK. recipe upload successfuly.
        "400":
          description: Bad Request. one of the argument is not specified.
      security:
      - cookieAuth: []
  /addToFavorites:
    post:
      tags:
      - profile
      description: mark recipe as a user's favorite.
      requestBody:
        description: New recipe that the user want to upload.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/body_2'
        required: true
      responses:
        "201":
          description: recipe added seccessfuly.
        "400":
          description: one of the argument is not specified.
        "409":
          description: recipe already marked as favorite by the user.
      security:
      - cookieAuth: []
...
  /viewMyRecipes:
    get:
      tags:
      - profile
      summary: Get all reciped that the user created.
      description: Get all recipes that the user created since registered.
      responses:
        "200":
          description: OK. recipes on the way.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreviewRecipe'
        "204":
          description: Bad, No Content was found.
        "400":
          description: Bad Request.
      security:
      - cookieAuth: []
...
components:
  schemas:
    Instruction:
      type: object
      properties:
        stepNumber:
          type: integer
          example: 1
        stepDescription:
          type: string
          example: Insert the gahnun to the oven, and you done :).
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/extendedIngredient'
      description: Respresent step in recipe, contain step number and step descripton.
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: curry.
      description: Represnt Ingredient.
    extendedIngredient:
      type: object
      properties:
        ingredient:
          $ref: '#/components/schemas/Ingredient'
        amount:
          type: number
          example: 0.5
        Unit measure:
          type: string
          example: grams.
      description: Represnt Ingredient and amount.
    BasePreviewPecipe:
      required:
      - aggregateLikes
      - glutenFree
      - image
      - instructions
      - isBeenWatched
      - isFavorite
      - readyInMinutes
      - title
      - vegan
      - vegetarian
      properties:
        title:
          type: string
          example: gahnun
        readyInMinutes:
          type: integer
          example: 7
        aggregateLikes:
          type: integer
          example: 453
        vegan:
          type: boolean
          example: true
        vegetarian:
          type: boolean
          example: true
        glutenFree:
          type: boolean
          example: true
        isBeenWatched:
          type: boolean
          example: true
        isFavorite:
          type: boolean
          example: true
        image:
          type: string
          example: gahnun.jpg
      description: Base preview of recipe. containing some of the recipe information
        without id field  (use for upload a new recipe).
    PreviewRecipe:
      description: Preview of recipe. conatining some or recipe information with id
        field.
      allOf:
      - $ref: '#/components/schemas/BasePreviewPecipe'
      - required:
        - id
        type: object
        properties:
          id:
            type: integer
            example: 453
    BaseRecipe:
      description: Recipe object with Ingredients and amounts, how to make the recipe
        and number of dish this recipe provides (use for upload a new recipe).
      allOf:
      - $ref: '#/components/schemas/BasePreviewPecipe'
      - type: object
        properties:
          Ingredients:
            type: array
            items:
              $ref: '#/components/schemas/extendedIngredient'
          Instructions:
            type: array
            items:
              $ref: '#/components/schemas/Instruction'
          serving:
            type: integer
            example: 5
...
    body:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: Containing letter only.
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: Containing at least one letter and at least one numbers.
      example:
        username: admin
        password: a123456
    body_1:
      type: object
      properties:
        username:
          type: string
          example: Johnsmith1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        country:
          type: string
          example: Israel
        password:
          type: string
          example: password
        email:
          type: string
          example: johnsmith@gmail.com
        image:
          type: string
          example: https://cloudinary.com/image.jpg
    body_2:
      type: object
      properties:
        recipeID:
          type: integer
          example: 1
        spoonacular:
          type: boolean
          example: true
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie